refactor: 페이지네이션 로직 분리
